feat(descargas-censo): divide header y footer de los informes censo
