+++ modules/descargas/informe-censo/informe-censo.ts
@@ -3,6 +3,9 @@ import { loadImage, InformePDF, getAssetsURL } from '../model/informe.class';
 import { CensoBody } from './censo-body';
 import { Auth } from '../../../auth/auth.class';
 import { CensoMensualBody } from './censo-mensual-body';
+import { CensoHeader } from './censo-header';
+import { CensoFooter } from './censo-footer';
+import { CensoMensualHeader } from './censo-mensual-header';
 
 export class InformeCenso extends InformePDF {
     constructor(public tipoCenso, public req) {
@@ -20,11 +23,15 @@ export class InformeCenso extends InformePDF {
 
         if (this.tipoCenso === 'diario') {
             const fechaCenso = moment(this.req.body.fecha).format('DD/MM/YYYY');
-            this.body = new CensoBody(usuario, fechaCenso, organizacion, unidadOrganizativa, this.req.body.listadoCenso, this.req.body.resumenCenso);
+            this.header = new CensoHeader(fechaCenso, organizacion, unidadOrganizativa);
+            this.body = new CensoBody(this.req.body.listadoCenso, this.req.body.resumenCenso);
+            this.footer = new CensoFooter(usuario, organizacion);
         }
 
         if (this.tipoCenso === 'mensual') {
-            this.body = new CensoMensualBody(usuario, organizacion, unidadOrganizativa, this.req.body.fechaDesde, this.req.body.fechaHasta, this.req.body.listadoCenso, this.req.body.resumenCenso, this.req.body.datosCenso);
+            this.header = new CensoMensualHeader(organizacion, unidadOrganizativa, this.req.body.fechaDesde, this.req.body.fechaHasta);
+            this.body = new CensoMensualBody(this.req.body.listadoCenso, this.req.body.resumenCenso, this.req.body.datosCenso);
+            this.footer = new CensoFooter(usuario, organizacion);
         }
 
         // Obligatorio por ahora llamar al proccess de la clase abstracta
